ADR_VALUE and some paths
for SetupPaths.do, I added a couple new paths

I added ADR_Value to this folder.
And I also merged in the code that constructs the weight for the Value
Index.
In addition, I constructed a Value Index for local returns.
(Generating "$bbpath/LocalValueIndex.dta" We haven’t used this before,
so we’ll need to test it.  It has industry_sector and Ticker as
ValueIndex, but sets the market to “AR”.

Note: I did NOT change the code to calculate all of the other returns.

 Also, there are different firms in ValueIndex for the ADR and local
base.