Adding Factor Stuff back in
The factors were merged in with ValueIndex because ValueIndex enters
later than the other returns and factors.  This means that on line 457
in ThirdAnalysis.do, drop if SPX_ == . would remove ValueIndex.

We could do this a more elegant way but it involves writing a lot more
code.  Let’s discuss.