Updated paths, data sources, benchmark
I updated this for the new paths, and to use ThirdAnalysis. Because of
a bug/data issue about the MXAR, I changed the “example_sec” used to do
the CDS stuff to the ValueIndex_US. An alternative would be to rewrite
this to remove the need for the example_sec.